feat: 클래스를 분리하고 게임기 객체가 실행될 수 있도록 기능 구현
